Move the Flask app fixtures into test_unified_graphics
test_unified_graphics is the only module that needs access to the Flask
application, so I moved the fixtures that set that up in there, along
with the diag_zarr and diag_dataset fixtures (which are now only used in
that module). Next, I'd like to adjust the scopes on the fixtures so
we're not doing unnecessary work before and after each test, and so that
we can group some of our asserts into a class as separate test
functions.
